requirements.txt
mne>=1.2.0
mne-bids>=0.12.0
jupyterlab>=3.6.0
pandas>=1.5.0
pyyaml>=6.0
openai>=1.20.0
pydantic>=1.10.0
prefect>=2.10.0
manubot>=0.5.0
requests>=2.28.0
nbconvert>=7.0.0
nbformat>=5.8.0
pandoc>=2.3
matplotlib>=3.7.0
numpy>=1.22.0
scikit-learn>=1.2.0
autoreject>=0.4.0
pyqt5>=5.15.0